TST: Unify 2d/3d tests for incompressible flow
